requirements.txt
flask
python-dotenv
pandas
ibm-watson-machine-learning